Refactor: Remove 'name' and 'code' from Rate entities and DTOs. Add AbstractEntity/AbstractDTO.
