+++ backend/src/main/java/com/bowe/meetstudent/controllers/ProgramController.java
@@ -1,5 +1,6 @@
 package com.bowe.meetstudent.controllers;
 
+import com.bowe.meetstudent.services.ProgramService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,4 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(path = "/api/v1/programs")
 @io.swagger.v3.oas.annotations.tags.Tag(name = "12. Programs", description = "Endpoints for managing educational programs")
 public class ProgramController {
+
+    private final ProgramService programService;
+
 }
