+++ backend/src/main/java/com/bowe/meetstudent/dto/RateDTO.java
@@ -5,14 +5,16 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+@EqualsAndHashCode(callSuper = true)
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @SuperBuilder
-public class RateDTO extends BaseDTO {
+public class RateDTO extends AbstractDTO {
 
     @NotNull(message = "Note is required")
     @Min(value = 0, message = "Rating must be at least 0")
