+++ backend/src/main/java/com/bowe/meetstudent/entities/BaseEntity.java
@@ -1,13 +1,9 @@
 package com.bowe.meetstudent.entities;
 
-import com.bowe.meetstudent.utils.Utils;
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
-import org.hibernate.proxy.HibernateProxy;
-
-import java.time.LocalDateTime;
-import java.util.Objects;
 
 @Getter
 @Setter
@@ -15,15 +11,9 @@ import java.util.Objects;
 @AllArgsConstructor
 @SuperBuilder
 @MappedSuperclass
-@EqualsAndHashCode(onlyExplicitlyIncluded = true)
-@ToString(onlyExplicitlyIncluded = true)
-public class BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @ToString.Include
-    @EqualsAndHashCode.Include
-    private Integer id;
+@EqualsAndHashCode(callSuper = true, onlyExplicitlyIncluded = true)
+@ToString(callSuper = true, onlyExplicitlyIncluded = true)
+public class BaseEntity extends AbstractEntity {
 
     @Column(length = 5, unique = true)
     @ToString.Include
@@ -32,30 +22,4 @@ public class BaseEntity {
     @Column(length = 50)
     @ToString.Include
     private String name;
-
-    private LocalDateTime createdAt;
-
-    private LocalDateTime modifiedAt ;
-
-    private Integer createdBy;
-
-    private Integer modifiedBy;
-
-    @PrePersist
-    protected void onCreate() {
-        setCreatedAt(
-                Utils.dakarTimeZone()
-        );
-        setModifiedAt(
-                Utils.dakarTimeZone()
-        );
-    }
-
-    @PreUpdate
-    protected void onUpdate() {
-
-        this.setModifiedAt(
-                Utils.dakarTimeZone()
-        );
-    }
 }
