+++ backend/src/main/resources/db/migration/V6__remove_name_code_from_rates.sql
@@ -1,0 +1,11 @@
+-- Remove code and name columns from school_rates
+ALTER TABLE school_rates DROP COLUMN code;
+ALTER TABLE school_rates DROP COLUMN name;
+
+-- Remove code and name columns from program_rates
+ALTER TABLE program_rates DROP COLUMN code;
+ALTER TABLE program_rates DROP COLUMN name;
+
+-- Remove code and name columns from course_rates
+ALTER TABLE course_rates DROP COLUMN code;
+ALTER TABLE course_rates DROP COLUMN name;
